+++ userDefined.js
@@ -1,15 +1,14 @@
 function UserDefinedClass() {
   this.transactionDetails = {
-    keyID: 'nj9sbq46gvbuh',
-    secretID: 'rbe8jFYXI_D7_p5jUIIV9WXbHVjz78HutIdRUfpAJr8'
+    keyID: '-',
+    secretID: '-'
   };
   this.debug = false;
-  this.tradingVolume = '0.0005';
+  this.tradingVolume = '0.001';
   this.orderRandHistory = 80;
   this.ratioBuy = 0.80;
   this.ratioSell = 0.55;
-  this.hardLimitSellPositive = 500;
-  this.hardLimitSellNegative = 250;
+  this.hardLimit = 100;
 }
 
 var userDefinedModule = new UserDefinedClass();