+++ History.md
@@ -1,5 +1,10 @@
 #Revision History
 
+* 1.0.1
+* Fix issue-02 - Moving the buy state into the lookForBuy State
+* Fix issue-03 - Implement hard limit for the selling
+* Fix issue-04 - Pending orders should ignore anything higher then user defined limit and order by time
+
 # 0.0.0
 * Fix issue-01 - Creating and initializing the basic bot api